requirements.txt
argon2-cffi
Django
django-filter
django-tables2
djangorestframework
numpy
netCDF4
requests
# Uncomment the next line if the system sqlite is not recent enough for Django
# pysqlite3-binary
# Uncomment the following lines if you would like to be able to build the documentation
# sphinx
# sphinx-rtd-theme
# Uncomment the following lines if you would like to run pylint
# pylint
# pylint-django